backend/requirements.txt
# ============================================
# Core Framework
# ============================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# ============================================
# LangChain & LangGraph (AI Agents)
# ============================================
langchain==0.1.4
langchain-openai==0.0.5
langgraph==0.0.20
langchain-community==0.0.16

# ============================================
# AI Model Integration
# ============================================
openai==1.10.0
httpx==0.26.0

# ============================================
# Technical Analysis
# ============================================
TA-Lib==0.4.28
pandas==2.2.0
numpy==1.26.3
scipy==1.12.0

# ============================================
# WebSockets
# ============================================
python-socketio==5.11.0
python-binance==1.0.19
websocket-client==1.7.0
websockets==12.0

# ============================================
# Database & Caching
# ============================================
motor==3.3.2
pymongo==4.6.1
redis==5.0.1

# ============================================
# Authentication & Security
# ============================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
supabase==2.3.0
pyjwt==2.8.0

# ============================================
# HTTP & API
# ============================================
requests==2.31.0
aiohttp==3.9.1

# ============================================
# Utilities
# ============================================
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2024.1

# ============================================
# Logging & Monitoring
# ============================================
structlog==24.1.0
# sentry-sdk==1.40.0  # Optional

# ============================================
# Testing (Development)
# ============================================
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# ============================================
# Code Quality (Development)
# ============================================
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
